nudb: Use buffer instead of buffers
