+++ src/beast/beast/nudb/detail/stream.h
@@ -32,63 +32,6 @@ namespace beast {
 namespace nudb {
 namespace detail {
 
-// Simple growable memory buffer
-class buffer
-{
-private:
-    std::size_t size_ = 0;
-    std::unique_ptr<std::uint8_t[]> buf_;
-
-public:
-    buffer() = default;
-    buffer (buffer const&) = delete;
-    buffer& operator= (buffer const&) = delete;
-
-    explicit
-    buffer (std::size_t n)
-        : size_ (n)
-        , buf_ (new std::uint8_t[n])
-    {
-    }
-
-    buffer (buffer&& other)
-        : size_ (other.size_)
-        , buf_ (std::move(other.buf_))
-    {
-        other.size_ = 0;
-    }
-
-    buffer& operator= (buffer&& other)
-    {
-        size_ = other.size_;
-        buf_ = std::move(other.buf_);
-        other.size_ = 0;
-        return *this;
-    }
-
-    std::size_t
-    size() const
-    {
-        return size_;
-    }
-
-    std::uint8_t*
-    get() const
-    {
-        return buf_.get();
-    }
-
-    void
-    reserve (std::size_t n)
-    {
-        if (size_ < n)
-            buf_.reset (new std::uint8_t[n]);
-        size_ = n;
-    }
-};
-
-//------------------------------------------------------------------------------
-
 // Input stream from bytes
 template <class = void>
 class istream_t
